+++ typescript/packages/plugins/jupiter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @goat-sdk/plugin-jupiter
 
+## 0.2.4
+
+### Patch Changes
+
+- de8533d: Fix transaction instructions
+- Updated dependencies [de8533d]
+  - @goat-sdk/wallet-solana@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes